+++ src/main/java/com/imsweb/algorithms/AlgorithmField.java
@@ -3,6 +3,9 @@
  */
 package com.imsweb.algorithms;
 
+/**
+ * Abstraction of an "field" that is used either as input or output (or both) for an given algorithm.
+ */
 public class AlgorithmField {
 
     public static AlgorithmField of(String id, Integer number, Integer length) {
@@ -13,8 +16,13 @@ public class AlgorithmField {
         return field;
     }
 
+    // field ID (for standard fields, it should be the NAACCR XML ID)
     private String _id;
+
+    // NAACCR Number (this will be available only for standard fields)
     private Integer _number;
+
+    // field length
     private Integer _length;
 
     public String getId() {