+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 **Changes in version 1.22 (not released yet)**
 
 - Changed internal initialization of NHIA and NAPIIA algorithms to make them thread-safe.
+- Changed default value of NAPIIA from an empty string to null.
 - Deprecated some methods using layout properties.
 - Added new framework to standardize the algorithms and allow them to be registered and executed via a factory.