+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Changed internal initialization of NHIA and NAPIIA algorithms to make them thread-safe.
 - Changed default value of NAPIIA from an empty string to null.
+- Added missing code SVK (for Slovakia) to the non-hispanic list based on birth place for NHIA.
 - Deprecated some methods using layout properties.
 - Added new framework to standardize the algorithms and allow them to be registered and executed via a factory.